+++ src/app.js
@@ -10,7 +10,7 @@ var corsOptions = {
   origin: "http://localhost:8081"
 };
 
-const db = require("./src/models");
+const db = require("./models");
 
 db.sequelize.sync();
 
@@ -22,12 +22,9 @@ app.use(bodyParser.json());
 // parse requests of content-type - application/x-www-form-urlencoded
 app.use(bodyParser.urlencoded({ extended: true }));
 
-require('./src/routes/auth')(app);
-require('./src/routes/user')(app);
-require('./src/routes/book')(app);
+require('./api/routes/auth')(app);
+require('./api/routes/user')(app);
+require('./api/routes/book')(app);
 
-// set port, listen for requests
-const PORT = process.env.PORT || 8080;
-app.listen(PORT, () => {
-  console.log(`Server is running on port ${PORT}.`);
-});
+module.exports = app;
+
