+++ src/slack-bot.js
@@ -1,5 +1,5 @@
 const { App } = require("@slack/bolt");
-const AddBook = require("./src/slack-bot/modals/AddBook");
+const AddBook = require("./slack-bot/modals/AddBook");
 const bcrypt = require("bcryptjs");
 
 require("dotenv").config();
@@ -11,10 +11,13 @@ const app = new App({
   appToken: process.env.APP_TOKEN
 });
 
-const db = require("./src/models");
+const db = require("./models");
 const Book = db.books;
 const User = db.users;
-db.sequelize.sync();
+// db.sequelize.sync();
+db.sequelize.sync({ force: true }).then(() => {
+  console.log("Drop and re-sync db.");
+});
 
 app.event('member_joined_channel', async ({ event, client, context }) => {
   try {
@@ -22,6 +25,7 @@ app.event('member_joined_channel', async ({ event, client, context }) => {
       channel: event.channel,
       text: 'hey budy, welcome!'
     });
+    console.log(event.user);
     resolveUser(client, event.user);
   } catch (error){
     console.error(error);
